requirements.txt
torch_geometric==1.4.3
numpy==1.18.4
torch_scatter==2.0.4
torch_sparse==0.6.3
scikit_learn==0.23.0
